+++ backend/modules/notifications/serializers.py
@@ -5,9 +5,9 @@ from modules.investors.models import Investor
 from modules.notifications.models import PriceAlert, PushSubscription
 
 
-class NotificationSerializer(serializers.ModelSerializer):
+class NotificationConfigSerializer(serializers.ModelSerializer):
     class Meta:
-        abstract = True
+        model = NotificationConfig
         fields = [
             "id",
             "is_email",
@@ -32,87 +32,6 @@ class NotificationCreateSerializer(serializers.ModelSerializer):
     push_subscription = PushNotificationSerializer(required=False)
 
     class Meta:
-        abstract = True
+        model = NotificationConfig
         fields = ["is_email", "is_push", "is_websocket", "push_subscription"]
 
-
-class PriceAlertSerializer(NotificationSerializer):
-    instrument_name = serializers.CharField(source="instrument.name", read_only=True)
-    instrument_ticker = serializers.CharField(
-        source="instrument.ticker", read_only=True
-    )
-
-    class Meta(NotificationSerializer.Meta):
-        model = PriceAlert
-        fields = NotificationSerializer.Meta.fields + [
-            "instrument_name",
-            "instrument_ticker",
-            "threshold_type",
-            "threshold_value",
-        ]
-
-
-class PriceAlertCreateSerializer(NotificationCreateSerializer):
-    instrument_ticker = serializers.CharField(write_only=True)
-
-    class Meta(NotificationCreateSerializer.Meta):
-        model = PriceAlert
-        fields = NotificationCreateSerializer.Meta.fields + [
-            "instrument_ticker",
-            "threshold_type",
-            "threshold_value",
-        ]
-
-    def validate_instrument_ticker(self, value):
-        if not Instrument.objects.filter(ticker=value).exists():
-            raise serializers.ValidationError(
-                f"Instrument with ticker '{value}' does not exist."
-            )
-        return value
-
-    def create(self, validated_data):
-        push_subscription = validated_data.pop("push_subscription", None)
-        instrument_ticker = validated_data.pop("instrument_ticker")
-
-        # Get the instrument
-        try:
-            instrument = Instrument.objects.get(ticker=instrument_ticker)
-        except Instrument.DoesNotExist as err:
-            raise serializers.ValidationError(
-                f"Instrument with ticker '{instrument_ticker}' does not exist."
-            ) from err
-
-        # Get the investor
-        request = self.context.get("request")
-        if not request or not request.user:
-            raise serializers.ValidationError("User must be authenticated.")
-
-        try:
-            investor = Investor.objects.get(clerk_id=request.user.id)
-        except Investor.DoesNotExist as err:
-            raise serializers.ValidationError("Investor profile not found.") from err
-
-        # Create or update the PriceAlert
-        price_alert, _ = PriceAlert.objects.update_or_create(
-            investor=investor,
-            instrument=instrument,
-            threshold_type=validated_data["threshold_type"],
-            threshold_value=validated_data["threshold_value"],
-            defaults={
-                "is_email": validated_data["is_email"],
-                "is_push": validated_data["is_push"],
-                "is_websocket": validated_data.get("is_websocket", True),
-                "is_active": validated_data.get("is_active", True),
-            },
-        )
-
-        # Create PushSubscription
-        if push_subscription:
-            PushSubscription.objects.update_or_create(
-                investor=investor,
-                endpoint=push_subscription["endpoint"],
-                p256dh=push_subscription["p256dh"],
-                auth=push_subscription["auth"],
-            )
-
-        return price_alert
